+++ bpf/bpf_lxc.c
@@ -566,7 +566,7 @@ ct_recreate4:
 
 #ifdef ENABLE_NODEPORT
 		/* This handles reply traffic for the case where the nodeport EP
-		 * is local to the node. We'll redirect to bpf_netdev egress to
+		 * is local to the node. We'll redirect to bpf_host egress to
 		 * perform the reverse DNAT.
 		 */
 		if (ct_state.node_port) {
@@ -945,7 +945,7 @@ int tail_ipv6_policy(struct __ctx_buff *ctx)
 		return send_drop_notify(ctx, src_label, SECLABEL, LXC_ID,
 					ret, CTX_ACT_DROP, METRIC_INGRESS);
 
-	// Store meta: essential for proxy ingress, see bpf_hostdev_ingress.c
+	// Store meta: essential for proxy ingress, see bpf_host.c
 	ctx_store_meta(ctx, 0, ctx->mark);
 	return ret;
 }
@@ -1165,7 +1165,7 @@ int tail_ipv4_policy(struct __ctx_buff *ctx)
 		return send_drop_notify(ctx, src_label, SECLABEL, LXC_ID,
 					ret, CTX_ACT_DROP, METRIC_INGRESS);
 
-	// Store meta: essential for proxy ingress, see bpf_hostdev_ingress.c
+	// Store meta: essential for proxy ingress, see bpf_host.c
 	ctx_store_meta(ctx, 0, ctx->mark);
 	return ret;
 }
