+++ pkg/common/const.go
@@ -33,7 +33,7 @@ const (
 	// CHeaderFileName is the name of the C header file for BPF programs for a
 	// particular endpoint.
 	CHeaderFileName = "lxc_config.h"
-	// NetdevHeaderFileName is the name of the header file used for bpf_netdev.c and bpf_overlay.c.
+	// NetdevHeaderFileName is the name of the header file used for bpf_host.c and bpf_overlay.c.
 	NetdevHeaderFileName = "netdev_config.h"
 	// PreFilterHeaderFileName is the name of the header file used for bpf_xdp.c.
 	PreFilterHeaderFileName = "filter_config.h"
